Increase step timeout from 20 mins to 2 hours.
The steps timeout when there is no output for 20 minutes. Because we
capture the stderr for the emulator tests (e.g. boot test), it can take
almost an hour until there is any output. For now, let's increase the
timeout. The more long-term solution is to print the output in real
time and then looking at it. We need to determine with the chrome infra
folks how to do this with recipes.

Change-Id: I2a1361d5b568b019a918a03c840ef82adfd815bc