Add debugging statements for mail notifications.
The mail notifications were not being sent successfully. This change was
made on the master to debug, but then after the restart the mail
notifications began being sent. Hopefully, they will continue to work,
but if they ever stop, these log statements will be useful in
debugging. This also sets the mac poller to 300 seconds instead of 30.

Change-Id: Ie0659e21f68fade6ff55d064fa2a17ff6710bdcc